# sevfruit-supplier-service/api/api-spec.yaml
...
openapi: 3.1.0
info:
  title: Сервис учёта поставок
  description: |
    API сервиса учёта поставок.
    Используется для внесения информации о поставщиках, продукции, ценах на продукцию на каждый период поставок и поставках продукции.
    Предусмотрено создание отчёта о поставках продукции поставщиками в заданном периоде.
    
    ## Авторизация  
    Для определения прав доступа к методам API используется авторизация API key. Ключ доступа передаётся в заголовке *Api-Key*.  
    
    ## Ошибки  
    API использует стандартные HTTP коды для сообщения об успешности или не успешности вызова API. В случае неуспеха, теле ответа
    содержится JSON с информацией об ошибке.
  version: '1.0'
  license:
    name: Unlicense
    url: https://opensource.org/license/unlicense/
servers:
  - url: http://localhost:8080
    description: Локальный сервер
security:
  - apikey: []
...
paths:
  /supplier:
    get:
      operationId: supplierGetAll
      summary: Список всех поставщиков
      description: Возвращает список всех поставщиков
      tags:
      - Supplier
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
              example:
              - id: 1
                name: S1
              - id: 2
                name: S2
              - id: 3
                name: S3
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /supplier
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /supplier
    post:
      operationId: supplierAdd
      summary: Добавить нового поставщика
      description: Добавляет нового поставщика
      tags:
      - Supplier
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Supplier"
            example:
              name: S4
      responses:
        "201":
          description: OK
          headers:
            Location:
              description: URL нового поставщика
              schema:
                type: string
                format: uri
              example: http://localhost:8080/supplier/4
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
              example:
                id: 4
                name: S4
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /supplier
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /supplier
  /supplier/{supplier_id}:
    get:
      operationId: supplierGet
      summary: Заданный поставщик
      description: Возвращает заданного поставщика
      tags:
      - Supplier
      parameters:
        - name: supplier_id
          in: path
          description: ID поставщика
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content: 
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
              example:
                id: 1
                name: "S1"
        "401":
          description: Запрос не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 401
                error: Unauthorized
                message: Full authentication is required to access this resource
                path: /supplier/1
        "404":
          description: Поставщик не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 404
                error: Not Found
                message: Поставщик 11 не найден
                path: /supplier/11
        "500":
          description: Ошибка сервера
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                timestamp: 2023-12-15T14:42:03.723+00:00
                status: 500
                error: Internal Server Error
                message: could not execute SQL statement
                path: /supplier/1
...
components:
  schemas:
    Supplier:
      type: object
      title: Поставщик
      description: Описание поставщика
      properties:
        id:
          type: integer
          format: int32
          description: ID поставщика
          readOnly: true
        name:
          type: string
          description: Наименование поставщика, уникальное для данного сервиса
      required:
      - name
...
    ErrorResponse:
      type: object
      title: Ошибка
      description: Описание ошибки
      properties:
        timestamp:
          type: string
          description: Дата и время
          readOnly: true
        status:
          type: integer
          description: HTTP код ошибки
          readOnly: true
        error:
          type: string
          description: Текстовое описание HTTP кода ошибки
          readOnly: true
        message:
          type: string
          description: Сообщение об ошибке
          readOnly: true
        path: 
          type: string
          description: Путь запроса
          readOnly: true
      required:
      - timestamp
      - status
      - error
  securitySchemes:
    apikey:
      type: apiKey
      in: header
      name: Api-Key
      description: Авторизация с помощью API key